requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.6.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-nginx==1.5.7
sqlparse==0.5.3
djangorestframework-simplejwt==5.3.1
channels==4.0.0
daphne==4.0.0
channels_redis==4.1.0
Pillow==12.0.0
whitenoise==6.8.2
django-import-export==3.3.1
